+++ src/filesystem.h
@@ -1,6 +1,10 @@
 #ifndef __HIBUS_BUSYBOX_FS__
 #define __HIBUS_BUSYBOX_FS__
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 char * listDirectory(hibus_conn* conn, const char* from_endpoint, const char* to_method, const char* method_param, int *err_code);
 char * removeFile(hibus_conn* conn, const char* from_endpoint, const char* to_method, const char* method_param, int *err_code);
 char * removeDirectory(hibus_conn* conn, const char* from_endpoint, const char* to_method, const char* method_param, int *err_code);
@@ -10,5 +14,9 @@ char * touchFile(hibus_conn* conn, const char* from_endpoint, const char* to_met
 
 void fs_register(hibus_conn *hibus_context);
 void fs_revoke(hibus_conn *hibus_context);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif  // __HIBUS_BUSYBOX_FS__
