+++ test/CMakeLists.txt
@@ -3,13 +3,12 @@ project(test)
 
 FIND_LIBRARY(libhibox NAMES hibox)
 FIND_LIBRARY(libhibus NAMES hibus)
-FIND_LIBRARY(libhibus-busybox NAMES hibus-busybox)
 
 INCLUDE_DIRECTORIES(BEFORE ${BUSYBOX_PUBLIC_HEADER_DIR})
+LINK_DIRECTORIES(${BUSYBOX_PUBLIC_LIB_DIR})
 
 AUX_SOURCE_DIRECTORY(./ DIR_SRCS)
 
 # make library
 add_executable(test ${DIR_SRCS})
-#TARGET_LINK_LIBRARIES(test pthread dl ${libhibox} ${libhibus} ${hibus-busybox})
-TARGET_LINK_LIBRARIES(test pthread ${libhibox} ${libhibus} ${libhibus-busybox} dl)
+TARGET_LINK_LIBRARIES(test pthread ${libhibox} ${libhibus} hibus-busybox dl)
